Redesign settings: split monolithic page into separate pages accessible from user dropdown
Break the single overloaded settings page into 5 focused pages, each accessible
from the sidebar user dropdown menu:

- General (/settings) - org name only
- Integrations (/integrations) - GitHub, Sentry, Linear connections
- Agent (/agent) - credentials, execution behavior, confidence thresholds
- Prioritization (/prioritization) - PM agent, product context, priority weights
- Team (/team) - unchanged

This removes the collapsible "Advanced Settings" junk drawer, eliminates the
unnecessary Agent Setup redirect card, and gives each domain its own page with
a dedicated Save button.
