+++ frontend/src/app/(dashboard)/settings/layout.tsx
@@ -1,52 +1,18 @@
 "use client";
 
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { cn } from "@/lib/utils";
 import { PageHeader } from "@/components/page-header";
-
-const settingsTabs = [
-  { label: "General", href: "/settings" },
-];
 
 export default function SettingsLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const pathname = usePathname();
-
   return (
     <div className="space-y-6">
       <PageHeader
-        title="Organization Settings"
-        description="Manage your organization and integrations."
+        title="General Settings"
+        description="Manage your organization."
       />
-      <nav className="flex gap-1 border-b border-border">
-        {settingsTabs.map((tab) => {
-          const isActive =
-            tab.href === "/settings"
-              ? pathname === "/settings"
-              : pathname.startsWith(tab.href);
-          return (
-            <Link
-              key={tab.href}
-              href={tab.href}
-              className={cn(
-                "relative px-3 py-2 text-sm font-medium transition-colors",
-                isActive
-                  ? "text-foreground"
-                  : "text-muted-foreground hover:text-foreground"
-              )}
-            >
-              {tab.label}
-              {isActive && (
-                <span className="absolute inset-x-0 -bottom-px h-0.5 bg-foreground" />
-              )}
-            </Link>
-          );
-        })}
-      </nav>
       {children}
     </div>
   );
